fix(ios): widget review remediations (timeline policy, privacy, shared controls, dead code)
