+++ clients/ios/App/App/MyViewController.swift
@@ -6,8 +6,9 @@ import WebKit
 ///
 /// 1. Registers `NativeAuthPlugin`, `NativeBiometricPlugin`,
 ///    `VoiceAudioSessionPlugin`, `VoiceLiveActivityPlugin`,
-///    `ApnsEnvironmentPlugin`, `SelfHostedServersPlugin`, and
-///    `RecentChatsPlugin` as local plugin instances at bridge init time.
+///    `ApnsEnvironmentPlugin`, `SelfHostedServersPlugin`,
+///    `RecentChatsPlugin`, and `WidgetSnapshotPlugin` as local plugin
+///    instances at bridge init time.
 ///    These plugins live inside the App target (no SPM module) so the bridge
 ///    won't discover them automatically.
 ///
